[CHANGED] selftests: ovpn: add notification parsing and matching
Changelog:
 - added check_requirements.py and all the other files to TEST_FILES in
   the Makefile
 - changed the skip message to include a capitalized SKIP as requested
   by Jakub

To verify that netlink notifications are correctly emitted and contain
the expected fields, this commit uses the tools/net/ynl/pyynl/cli.py
script to create multicast listeners. These listeners record the
captured notifications to a JSON file, which is later compared to the
expected output.

Since this change introduces additional dependencies (jq, pyyaml,
jsonschema), the tests are configured to check for their presence and
conditionally skip the notification check if they are missing.

Signed-off-by: Ralf Lici <ralf@mandelbit.com>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>