[CHANGED] selftests: ovpn: check a/symmetric peer-id
Changelog:
 - fixed an off-by-one argument count bug in CMD_NEW_PEER

Extend the base test to verify that the correct peer-id is set in data
packet headers. This is done by capturing ping packets with ngrep during
the initial exchange and matching the first portion of the header
against the expected sequence for every connection.

Assume asymmetric peer-id as the default, but also add tests with
symmetric peer-id to validate legacy topology.

Cc: Shuah Khan <shuah@kernel.org>
Signed-off-by: Ralf Lici <ralf@mandelbit.com>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>